Move Batch Delete functions out
